+++ src/main/java/com/example/kkm/user/auth/service/UserAuthService.java
@@ -9,8 +9,8 @@ import com.example.kkm.user.auth.repository.UserRepository;
 import com.example.kkm.user.auth.security.TokenProvider;
 import com.example.kkm.user.domain.entity.Password;
 import com.example.kkm.user.domain.entity.User;
-import com.example.kkm.user.profile.service.domain.model.Profile;
-import com.example.kkm.user.profile.service.domain.repository.ProfileRepository;
+import com.example.kkm.user.profile.domain.model.Profile;
+import com.example.kkm.user.profile.domain.repository.ProfileRepository;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
